pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gmail-inbox-organizer"
version = "0.1.0"
description = "A TUI application to analyze Gmail inbox, show sender frequency, and export data"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
authors = [
    {name = "User"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Communications :: Email",
    "Topic :: Office/Business",
]
dependencies = [
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-api-python-client>=2.0.0",
    "textual>=0.41.0",
    "rich>=13.0.0",
]

[project.scripts]
inbox-organizer = "main:main"

[tool.setuptools]
py-modules = ["main", "auth", "gmail_analyzer", "tui"]
